Split mesh creation functions
